QA Round 2 fixes: configurable Worker endpoint, secret auth, no upstream leaks, Int64 totals
[C1] QuotaStore reads endpoint from Info.plist (QuotaEndpointURL key);
nil endpoint disables QuotaCardView entirely instead of showing fake errors.
[C2] Differential bandwidth math switched to Int64; first tick uses
session-average instead of 0 B/s for immediate UI feedback.
[C3] Worker no longer echoes Vultr response bodies to clients;
CORS Origin now configurable via env.
[H1] Removed await task.value from ping() — dedup pattern now actually
fires-and-forgets.
[H2] QuotaCardView.onAppear only refreshes when snapshot is nil and
not already fetching.
[H3] Poller switched from Task.detached to Task to inherit actor;
weak-self mid-loop now breaks cleanly.
[H5] Worker supports optional CLIENT_SECRET header — iOS sends matching
QuotaClientSecret from Info.plist if configured.
[M1] Removed stale "from /traffic" comment (now from /connections).

Known tech debt: dual StatsSnapshot files in App/ and Tunnel/ remain
duplicated; consolidation deferred to a project.yml shared-sources
restructure in a future round.

+++ App/Components/StatsCardView.swift
@@ -180,18 +180,18 @@ struct StatsCardView: View {
 
     // MARK: - Formatting
 
-    private func formatSpeed(_ bps: Int) -> String {
+    private func formatSpeed(_ bps: Int64) -> String {
         let bcf = ByteCountFormatter()
         bcf.countStyle = .binary
         bcf.allowedUnits = [.useKB, .useMB, .useGB]
-        return bcf.string(fromByteCount: Int64(bps)) + "/s"
+        return bcf.string(fromByteCount: bps) + "/s"
     }
 
-    private func formatBytes(_ bytes: Int) -> String {
+    private func formatBytes(_ bytes: Int64) -> String {
         let bcf = ByteCountFormatter()
         bcf.countStyle = .binary
         bcf.allowedUnits = [.useKB, .useMB, .useGB]
-        return bcf.string(fromByteCount: Int64(bytes))
+        return bcf.string(fromByteCount: bytes)
     }
 
     private func formatDuration(_ secs: TimeInterval) -> String {
